refactor: inline download dialog HTML assertions in tests
Remove downloadAdDialogExpectations.ts and assert popover markup directly in unit and E2E tests. Revert the .gitignore PDF entry and drop an extra blank line in app/client.ts.
